Add files via upload
In this code, we see that how Object-Oriented Programming language (Python) is having encapsulation in it.
Encapsulation: It is defined as the wrapping up of data under a single unit. It is the mechanism that binds together code and the data it manipulates. Another way to think about encapsulation is, it is a protective shield that prevents the data from being accessed by the code outside this shield.
In this code, we have private members and public members. As we know that private members can only be accessed by public members of the same class. So here the same thing is done, an object of the class has been created and the public members of the class are being called and the code is running.